+++ python/mlc_chat/utils.py
@@ -1,53 +1,4 @@
 """Common utils used in Python runtime for MLC chat."""
-import ctypes
-import os
-import sys
 
 import tvm
 import tvm._ffi.base
-
-from enum import Enum
-from . import libinfo
-
-
-def _load_mlc_llm_lib():
-    """Load mlc llm lib"""
-    if sys.platform.startswith("win32") and sys.version_info >= (3, 8):
-        for path in libinfo.get_dll_directories():
-            os.add_dll_directory(path)
-    lib_name = "mlc_llm" if tvm._ffi.base._RUNTIME_ONLY else "mlc_llm_module"
-    lib_path = libinfo.find_lib_path(lib_name, optional=False)
-    return ctypes.CDLL(lib_path[0]), lib_path[0]
-
-
-# only load once here
-if os.environ.get("SKIP_LOADING_MLCLLM_SO", "0") == "0":
-    _LIB, _LIB_PATH = _load_mlc_llm_lib()
-
-
-def quantization_keys():
-    return [
-        "q3f16_0",
-        "q4f16_0",
-        "q4f16_1",
-        "q4f32_0",
-        "q8f16_0",
-        "q0f16",
-        "q0f32",
-    ]
-
-
-class PlaceInPrompt(Enum):
-    """The place of an input message in a prompt."""
-
-    # The input message should have role names and corresponding seperators appended both prior to it and after it,
-    # making it a complete prompt.
-    All = 0
-    # The input message is only the beginning part of a prompt, no role name and separator should be appended after
-    # the message since there will be future messages appended after the message.
-    Begin = 1
-    # The input message is in the middle of a prompt, nothing should be appended before or after the message.
-    Middle = 2
-    # The input message is the ending part of a prompt, no role name and separator should be appended prior to it
-    # since the message is concatenated to some prior messages.
-    End = 3
